Move rules to lit-analyzer/src/rules
