+++ packages/lit-analyzer/src/analyze/default-lit-analyzer-context.ts
@@ -4,15 +4,15 @@ import * as tsServer from "typescript/lib/tsserverlibrary";
 import { analyzeComponents, analyzeLibDomHtmlElement } from "web-component-analyzer";
 import { getBuiltInHtmlCollection } from "./data/get-built-in-html-collection";
 import { getUserConfigHtmlCollection } from "./data/get-user-config-html-collection";
-import noBooleanInAttributeBindingRule from "./document-analyzer/html/diagnostic/rules/assignment/no-boolean-in-attribute-binding";
-import noComplexAttributeBindingRule from "./document-analyzer/html/diagnostic/rules/assignment/no-complex-attribute-binding";
-import noExpressionlessPropertyBindingRule from "./document-analyzer/html/diagnostic/rules/assignment/no-expressionless-property-binding";
-import noIncompatibleTypeBindingRule from "./document-analyzer/html/diagnostic/rules/assignment/no-incompatible-type-binding";
-import noInvalidDirectiveBindingRule from "./document-analyzer/html/diagnostic/rules/assignment/no-invalid-directive-binding";
-import noNoncallableEventBindingRule from "./document-analyzer/html/diagnostic/rules/assignment/no-noncallable-event-binding";
-import noNullableAttributeBindingRule from "./document-analyzer/html/diagnostic/rules/assignment/no-nullable-attribute-binding";
-import noUnknownSlotRule from "./document-analyzer/html/diagnostic/rules/assignment/no-unknown-slot";
-import { RuleModule } from "./document-analyzer/html/diagnostic/rules/rule-module";
+import noBooleanInAttributeBindingRule from "../rules/no-boolean-in-attribute-binding";
+import noComplexAttributeBindingRule from "../rules/no-complex-attribute-binding";
+import noExpressionlessPropertyBindingRule from "../rules/no-expressionless-property-binding";
+import noIncompatibleTypeBindingRule from "../rules/no-incompatible-type-binding";
+import noInvalidDirectiveBindingRule from "../rules/no-invalid-directive-binding";
+import noNoncallableEventBindingRule from "../rules/no-noncallable-event-binding";
+import noNullableAttributeBindingRule from "../rules/no-nullable-attribute-binding";
+import noUnknownSlotRule from "../rules/no-unknown-slot";
+import { RuleModule } from "./types/rule-module";
 import { isRuleDisabled, LitAnalyzerConfig, makeConfig } from "./lit-analyzer-config";
 import { LitAnalyzerContext, LitPluginContextHandler } from "./lit-analyzer-context";
 import { DefaultLitAnalyzerLogger, LitAnalyzerLoggerLevel } from "./lit-analyzer-logger";
