--- packages/lit-analyzer/src/analyze/document-analyzer/html/diagnostic/rules/rule-module.ts
@@ -1,13 +1,0 @@
-import { LitAnalyzerRuleName } from "../../../../lit-analyzer-config";
-import { LitAnalyzerRequest } from "../../../../lit-analyzer-context";
-import { HtmlNodeAttrAssignment } from "../../../../types/html-node/html-node-attr-assignment-types";
-import { HtmlNodeAttr } from "../../../../types/html-node/html-node-attr-types";
-import { HtmlNode } from "../../../../types/html-node/html-node-types";
-import { LitHtmlDiagnostic } from "../../../../types/lit-diagnostic";
-
-export interface RuleModule {
-	name: LitAnalyzerRuleName;
-	visitHtmlNode?(node: HtmlNode, context: LitAnalyzerRequest): LitHtmlDiagnostic[] | void;
-	visitHtmlAttribute?(attribute: HtmlNodeAttr, context: LitAnalyzerRequest): LitHtmlDiagnostic[] | void;
-	visitHtmlAssignment?(assignment: HtmlNodeAttrAssignment, context: LitAnalyzerRequest): LitHtmlDiagnostic[] | void;
-}
