+++ packages/lit-analyzer/src/analyze/util/directive/is-lit-directive.ts
