+++ packages/lit-analyzer/src/analyze/util/type/is-assignable-in-boolean-binding.ts
@@ -1,8 +1,8 @@
 import { SimpleType, SimpleTypeKind, toTypeString } from "ts-simple-type";
-import { litDiagnosticRuleSeverity } from "../../../../../../lit-analyzer-config";
-import { LitAnalyzerRequest } from "../../../../../../lit-analyzer-context";
-import { HtmlNodeAttr } from "../../../../../../types/html-node/html-node-attr-types";
-import { LitHtmlDiagnostic, LitHtmlDiagnosticKind } from "../../../../../../types/lit-diagnostic";
+import { litDiagnosticRuleSeverity } from "../../lit-analyzer-config";
+import { LitAnalyzerRequest } from "../../lit-analyzer-context";
+import { HtmlNodeAttr } from "../../types/html-node/html-node-attr-types";
+import { LitHtmlDiagnostic, LitHtmlDiagnosticKind } from "../../types/lit-diagnostic";
 import { isAssignableToType } from "./is-assignable-to-type";
 
 export function isAssignableInBooleanBinding(
