+++ packages/lit-analyzer/src/analyze/util/type/remove-undefined-from-type.ts
