+++ packages/lit-analyzer/src/rules/no-incompatible-type-binding.ts
@@ -2,12 +2,12 @@ import {
 	LIT_HTML_BOOLEAN_ATTRIBUTE_MODIFIER,
 	LIT_HTML_EVENT_LISTENER_ATTRIBUTE_MODIFIER,
 	LIT_HTML_PROP_ATTRIBUTE_MODIFIER
-} from "../../../../../constants";
-import { RuleModule } from "../rule-module";
-import { extractBindingTypes } from "./util/extract-binding-types";
-import { isAssignableInAttributeBinding } from "./util/is-assignable-in-attribute-binding";
-import { isAssignableInBooleanBinding } from "./util/is-assignable-in-boolean-binding";
-import { isAssignableInPropertyBinding } from "./util/is-assignable-in-property-binding";
+} from "../analyze/constants";
+import { RuleModule } from "../analyze/types/rule-module";
+import { extractBindingTypes } from "../analyze/util/type/extract-binding-types";
+import { isAssignableInAttributeBinding } from "../analyze/util/type/is-assignable-in-attribute-binding";
+import { isAssignableInBooleanBinding } from "../analyze/util/type/is-assignable-in-boolean-binding";
+import { isAssignableInPropertyBinding } from "../analyze/util/type/is-assignable-in-property-binding";
 
 const rule: RuleModule = {
 	name: "no-incompatible-type-binding",
