+++ packages/lit-analyzer/src/rules/no-unknown-slot.ts
@@ -1,8 +1,8 @@
-import { litDiagnosticRuleSeverity } from "../../../../../lit-analyzer-config";
-import { HtmlNodeAttrAssignmentKind } from "../../../../../types/html-node/html-node-attr-assignment-types";
-import { HtmlNodeAttrKind } from "../../../../../types/html-node/html-node-attr-types";
-import { LitHtmlDiagnosticKind } from "../../../../../types/lit-diagnostic";
-import { RuleModule } from "../rule-module";
+import { litDiagnosticRuleSeverity } from "../analyze/lit-analyzer-config";
+import { HtmlNodeAttrAssignmentKind } from "../analyze/types/html-node/html-node-attr-assignment-types";
+import { HtmlNodeAttrKind } from "../analyze/types/html-node/html-node-attr-types";
+import { LitHtmlDiagnosticKind } from "../analyze/types/lit-diagnostic";
+import { RuleModule } from "../analyze/types/rule-module";
 
 const rule: RuleModule = {
 	name: "no-unknown-slot",
